AVR/avr-yandex-speechkit-adapter/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
PyJWT[crypto]==2.10.1
cryptography==44.0.0
